feat(#541): ensure fresh db setup per request
