+++ .github/workflows/deploy-production.yml
@@ -46,6 +46,7 @@ jobs:
           PRIVATE_KEY: ${{ secrets.EVM_SC_OWNER_PRIVATE_KEY }}
           DEFAULT_DB_URL: ${{ secrets.POSTGRES_URL }}
           DEFAULT_DB_AUTHENTICATED_URL: ${{ secrets.POSTGRES_AUTHENTICATED_URL }}
+          DEFAULT_DB_ANONYMOUS_URL: ${{secrets.POSTGRES_ANONYMOUS_URL}}
 
       - name: Deploy Project Artifacts to Vercel
         run: vercel deploy --prebuilt --prod --token=${{ secrets.VERCEL_TOKEN }}
