+++ apps/web/src/app/api/v1/people/me/route.ts
@@ -1,0 +1,41 @@
+import { headers } from 'next/headers';
+import { NextResponse } from 'next/server';
+import { createPeopleService } from '@hypha-platform/core';
+
+export async function GET() {
+  const headersList = await headers();
+  const authToken = headersList.get('Authorization')?.split(' ')[1] || '';
+  try {
+    if (!authToken) {
+      return NextResponse.json(
+        { error: 'Unauthorized' },
+        {
+          status: 401,
+        },
+      );
+    }
+
+    console.debug('getUser', { authToken: authToken.split('.')[2] });
+    const peopleService = createPeopleService({ authToken });
+    const user = await peopleService.findMe();
+
+    if (!user) {
+      return NextResponse.json(
+        { error: 'User not found' },
+        {
+          status: 404,
+        },
+      );
+    }
+
+    return NextResponse.json({ ...user });
+  } catch (error) {
+    console.error('Error fetching current user:', error);
+    return NextResponse.json(
+      { error: 'Internal Server Error' },
+      {
+        status: 500,
+      },
+    );
+  }
+}
