+++ package.json
@@ -69,7 +69,7 @@
     "lucide-react": "^0.469.0",
     "mdx": "^0.3.1",
     "negotiator": "^1.0.0",
-    "next": "15.1.5",
+    "next": "15.2.1",
     "next-i18n-router": "^5.5.1",
     "next-mdx-remote": "^5.0.0",
     "next-themes": "^0.4.4",
